move config title to helper
