+++ test/vscode-ui/test/specs/nested-fastapi.spec.ts
@@ -5,11 +5,7 @@ import * as fs from "fs";
 import * as path from "path";
 import { fileURLToPath } from "url";
 import { dirname } from "path";
-import {
-  runShellScript,
-  switchToSubframe,
-  waitForInputFields,
-} from "../helpers.ts";
+import * as helper from "../helpers.ts";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -39,7 +35,7 @@ describe("Nested Fast API Deployment", () => {
   });
 
   it("can add deployment", async () => {
-    await switchToSubframe();
+    await helper.switchToSubframe();
     // initialize project via button
     const selectButton = (await $('[data-automation="select-deployment"]')).$(
       ".quick-pick-label",
@@ -143,14 +139,14 @@ describe("Nested Fast API Deployment", () => {
     await browser.keys("\uE007");
 
     // wait until the server responds
-    await waitForInputFields("The API key to be used");
+    await helper.waitForInputFields("The API key to be used");
 
     //set api key
     await input.setValue(apiKey);
     await browser.keys("\uE007");
 
     // wait for server validation
-    await waitForInputFields("Enter a Unique Nickname");
+    await helper.waitForInputFields("Enter a Unique Nickname");
 
     // set server name
     await input.setValue("my connect server");
@@ -158,14 +154,7 @@ describe("Nested Fast API Deployment", () => {
   });
 
   it("can check config", async () => {
-    const workbench = await browser.getWorkbench();
-    const openEditorTitles = await workbench
-      .getEditorView()
-      .getOpenEditorTitles();
-    const fileNamePattern = /^my fastapi app-.*$/;
-    const realFilename = openEditorTitles.find((title) =>
-      fileNamePattern.test(title),
-    );
+    const realFilename = await helper.getConfigTitle(/^my fastapi app-.*$/);
 
     const filePath = path.resolve(
       __dirname,
@@ -214,7 +203,7 @@ describe("Nested Fast API Deployment", () => {
       it("remove credentials", async () => {
         let scriptPath: string;
         scriptPath = "cd ../scripts && bash cleanup.bash";
-        await runShellScript(scriptPath);
+        await helper.runShellScript(scriptPath);
       });
     });
   });
